Separated timeline item menu into its own component
